Implement a simple in-memory sort.
This is currently a dreadful and incorrect version of `ExternalSort`
however it's a good start to the problem. I've implemented a basic
`TupleComparator` that compares the columns of the tuple from left to
right. It needs some further changes to allow it to sort based on the
order that the user specifies.

Signed-off-by: Chris Brown <xoebus@xoeb.us>